requirements.txt
python-telegram-bot==22.2
pandas==2.2.0
google-generativeai==0.3.2
flask==3.0.0
openai-whisper==20231117
pandas==2.2.0
ffmpeg-python==1.4
